% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_sparse_data.R
\name{gen_sparse_multivar_data}
\alias{gen_sparse_multivar_data}
\title{function to generate data with hierarchical sparsity}
\usage{
gen_sparse_multivar_data(
  nvars = 10L,
  noutcomes = 8L,
  nobs = 100L,
  nobs.test = 100L,
  outcome.grouping = rbind(c(1, 1, 1, 2, 2, 2, 2, 2), c(1, 1, 1, 2, 2, 3, 3, 3), c(1,
    1, 2, 3, 3, 4, 4, 5), c(1:8)),
  effect.size.similarity = c("sign", "mean", "none"),
  hier.sparsity.prob = 0.1,
  prop.zero.vars = 0.5,
  family = c("gaussian", "binomial"),
  sd = 1,
  snr = NULL,
  beta = NULL,
  tau = 10,
  x.covar = 0,
  y.covar = 0.5
)
}
\arguments{
\item{nvars}{number of variables}

\item{noutcomes}{number of outcomes}

\item{nobs}{number of observations per outcomes to simulate}

\item{nobs.test}{number of independent test observations per outcomes to simulate}

\item{outcome.grouping}{groups for outcomes -- either a vector of length \code{noutcomes} of indices of what groups each outcome belongs to
or a matrix with \code{noutcomes} columns, with each row as a different grouping of the outcomes -- this allows for potentially overlapping
groups of the outcomes}

\item{effect.size.similarity}{How should the effect sizes of a variable across the outcomes be related? \code{"sign"} for
the signs to match, \code{"mean"} for the coefficients to be the same plus some small uniform noise, or \code{"none"} for there
to be no relationship at all}

\item{hier.sparsity.prob}{probability that a group has its coefficients set to zero for any variable}

\item{prop.zero.vars}{proportion of all variables that will be zero across all outcomes}

\item{family}{family for the response variable}

\item{sd}{standard devation for gaussian simulations}

\item{beta}{a matrix of true beta values. If given, then no beta will be created and data will be simulated from the given beta}

\item{x.covar}{scalar, pairwise covariance term for covariates}

\item{y.covar}{scalar, pairwise covariance term for outcomes}

\item{effect.size.max}{maximum magnitude of the true effect sizes}
}
\description{
function to generate data with hierarchical sparsity
}
\examples{
set.seed(123)

dat.sim <- gen_sparse_multivar_data(nvars = 10L,
                   noutcomes = 8L,
                   nobs = 100L,
                   nobs.test = 100L,
                   outcome.grouping = rbind(c(1,1,1,2,2,2,2,2),
                                            c(1,1,1,2,2,3,3,3),
                                            c(1,1,2,3,3,4,4,5),
                                            c(1:8)))

x        <- dat.sim$x
x.test   <- dat.sim$x.test
y        <- dat.sim$y
y.test   <- dat.sim$y.test
beta     <- dat.sim$beta

\dontrun{

outcome_groups <- rbind(c(1,1,1,2,2,2,2,2),
                        c(1,1,1,2,2,3,3,3),
                        c(1,1,2,3,3,4,4,5))
                        
fit.adapt <- cv.groupFusedMulti(x, y,
                                nlambda        = 50,
                                lambda.fused = c(0.000005, 0.00001, 0.000025, 0.00005, 0.0001),
                                outcome.groups = outcome_groups,
                                gamma          = 0.25,
                                nfolds         = 5)

preds.a <- predict(fit.adapt, x.test, type = 'response')
                   
}


}
