% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_fused_multi.R
\name{groupFusedMulti}
\alias{groupFusedMulti}
\title{Fitting groupFusedMulti}
\usage{
groupFusedMulti(
  x,
  y,
  outcome.groups = NULL,
  family = c("gaussian", "binomial"),
  nlambda = 100L,
  nlambda.fused = 0,
  lambda = NULL,
  lambda.min.ratio = NULL,
  lambda.fused = NULL,
  use.alpha.param = TRUE,
  fuse.level = c("last_group", "first_group", "all"),
  fuse.none.group = TRUE,
  lambda.fused.min = ifelse(ncol(x) >= nrow(x), 0.001, 1e-05),
  lambda.fused.max.ratio = 1e-05,
  penalty.factor = NULL,
  group.weights = NULL,
  adaptive.lasso = FALSE,
  adaptive.fused = FALSE,
  gamma = 1,
  standardize = FALSE,
  intercept = TRUE,
  one.intercept = FALSE,
  rho = NULL,
  dynamic.rho = TRUE,
  maxit = 500L,
  abs.tol = 1e-05,
  rel.tol = 1e-05,
  irls.tol = 1e-05,
  irls.maxit = 100L,
  model.matrix = FALSE,
  ...
)
}
\arguments{
\item{x}{input matrix of dimension nobs by nvars. Each row is an observation,
each column corresponds to a covariate}

\item{y}{numeric response matrix with \code{nobs} rows and \code{noutcomes} number of columns (one outcome per column)}

\item{outcome.groups}{a vector of length equal to the number of columns in the outcome matrix \code{y}}

\item{family}{\code{"gaussian"} for least squares problems}

\item{nlambda}{The number of lambda values. Default is 100.}

\item{nlambda.fused}{The number of lambda values for the fused lasso. Default is 0, which means that the fused
lasso is not used.}

\item{lambda}{A user-specified sequence of lambda values. Left unspecified, the a sequence of lambda values is
automatically computed, ranging uniformly on the log scale over the relevant range of lambda values.}

\item{lambda.min.ratio}{Smallest value for lambda, as a fraction of \code{lambda.max}, the (data derived) entry
value (i.e. the smallest value for which all parameter estimates are zero). The default
depends on the sample size \code{nobs} relative to the number of variables \code{nvars}. If
\code{nobs > nvars}, the default is 0.0001, close to zero. If \code{nobs < nvars}, the default
is 0.01. A very small value of \code{lambda.min.ratio} can lead to a saturated fit
when \code{nobs < nvars}.}

\item{lambda.fused}{tuning parameter for fused lasso penalty}

\item{use.alpha.param}{Should the 'alpha' parameterization for the fused lasso be used? alpha=1 will correspond to no fused lasso and alpha=0 will be all fused lasso}

\item{fuse.none.group}{Should the coefficients of the none subpopulation be fused with coefficients for subpopulations with just one condition?}

\item{lambda.fused.min}{minimum value used for the fused lasso tuning parameter}

\item{lambda.fused.max.ratio}{scalar value that is the ratio of the largest to smallest values
for the fused lasso tuning parameter}

\item{penalty.factor}{vector of weights to be multiplied to the tuning parameter for the
group lasso penalty. A vector of length equal to the number of groups}

\item{group.weights}{A vector of values representing multiplicative factors by which each group's penalty is to
be multiplied. Often, this is a function (such as the square root) of the number of predictors in each group.
The default is to use the square root of group size for the group selection methods.}

\item{adaptive.lasso}{Flag indicating whether or not to use adaptive lasso weights. If set to \code{TRUE} and
\code{group.weights} is unspecified, then this will override \code{group.weights}. If a vector is supplied to group.weights,
then the \code{adaptive.lasso} weights will be multiplied by the \code{group.weights} vector}

\item{adaptive.fused}{Flag indicating whether or not to use adaptive fused lasso weights.}

\item{gamma}{power to raise the MLE estimated weights by for the adaptive lasso. defaults to 1}

\item{standardize}{Should the data be standardized? Defaults to \code{FALSE}.}

\item{intercept}{Should an intercept be fit? Defaults to \code{TRUE}}

\item{one.intercept}{Should a single intercept be fit for all outcome instead of one
for each outcome? Defaults to \code{FALSE}.}

\item{rho}{ADMM parameter. must be a strictly positive value. By default, an appropriate value is automatically chosen}

\item{dynamic.rho}{\code{TRUE}/\code{FALSE} indicating whether or not the rho value should be updated throughout the course of the ADMM iterations}

\item{maxit}{integer. Maximum number of ADMM iterations. Default is 500.}

\item{abs.tol}{absolute convergence tolerance for ADMM iterations for the relative dual and primal residuals.
Default is \code{10^{-5}}, which is typically adequate.}

\item{rel.tol}{relative convergence tolerance for ADMM iterations for the relative dual and primal residuals.
Default is \code{10^{-5}}, which is typically adequate.}

\item{irls.tol}{convergence tolerance for IRLS iterations. Only used if \code{family != "gaussian"}. Default is 10^{-5}.}

\item{irls.maxit}{integer. Maximum number of IRLS iterations. Only used if \code{family != "gaussian"}. Default is 100.}

\item{model.matrix}{logical flag. Should the design matrix used be returned?}

\item{...}{not used}
}
\value{
An object with S3 class "groupFusedMulti"
}
\description{
Fitting groupFusedMulti
}
\examples{
library(Matrix)

}
