% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_group_fused_multi.R
\name{predict.cv.groupFusedMulti}
\alias{predict.cv.groupFusedMulti}
\title{Prediction for Cross Validation Hierarchical Lasso Object}
\usage{
\method{predict}{cv.groupFusedMulti}(object, newx = NULL, s = c("lambda.min"), ...)
}
\arguments{
\item{object}{fitted cv.groupFusedMulti object}

\item{newx}{new matrix for predictions}

\item{s}{lambda value for the predictions. defaults to minimum}

\item{...}{parameters to be passed to predict.groupFusedMulti}
}
\value{
predictions or coefficients
}
\description{
Prediction for Cross Validation Hierarchical Lasso Object
}
